+++ examples/data-sources/meraki_sensor_network_relationships/data-source.tf
@@ -1,0 +1,3 @@
+data "meraki_sensor_network_relationships" "example" {
+  network_id = "L_123456"
+}